fix(MM-Script): Check MaxMind key and init config dir (#586)
Check if the user has set the MaxMind download key and if not give some hints on how to set it and where he can sign up to get one. Initialize the config directory so the user doesn't have to create it himself. The variable is now the same as in the Dockerfile.

+++ bin/get_mm.sh
@@ -1,12 +1,23 @@
-#!/bin/bash
+#!/usr/bin/env bash
 
-curl -o mm.tar.gz "https://download.maxmind.com/app/geoip_download?edition_id=GeoLite2-Country&license_key=${MM_DOWNLOAD_KEY}&suffix=tar.gz"
+if [ -z ${MAXMIND_LICENSE_KEY} ]; then
+  echo ""
+  echo "Provide a MaxMind license key to fetch the Geolite2 database files."
+  echo "export MAXMIND_LICENSE_KEY=<your-license-key>"
+  echo ""
+  echo "Sign-up for a free GeoLite2 account at https://dev.maxmind.com and create your license key."
+  echo ""
+  exit -1
+fi
+
+mkdir -p config
+curl -o mm.tar.gz "https://download.maxmind.com/app/geoip_download?edition_id=GeoLite2-Country&license_key=${MAXMIND_LICENSE_KEY}&suffix=tar.gz"
 tar -zxf mm.tar.gz
 mv GeoLite2-Country_*/GeoLite2-Country.mmdb config/GeoLite2-Country.mmdb
 rm mm.tar.gz
 rm -r GeoLite2-Country_*
 
-curl -o mm.tar.gz "https://download.maxmind.com/app/geoip_download?edition_id=GeoLite2-ASN&license_key=${MM_DOWNLOAD_KEY}&suffix=tar.gz"
+curl -o mm.tar.gz "https://download.maxmind.com/app/geoip_download?edition_id=GeoLite2-ASN&license_key=${MAXMIND_LICENSE_KEY}&suffix=tar.gz"
 tar -zxf mm.tar.gz
 mv GeoLite2-ASN_*/GeoLite2-ASN.mmdb config/GeoLite2-ASN.mmdb
 rm mm.tar.gz
